refactor: typescript migration
