+++ conf/conf.d/default.conf
@@ -1,6 +1,6 @@
 server {
-	listen 3000;
-	listen [::]:3000;
+	listen 80;
+	listen [::]:80;
 
 	location / {
 		root /home/ubuntu/STUDIO-EYE-WEB-MANAGEMENT/build;
