+++ package.json
@@ -52,5 +52,13 @@
   },
   "browser": {
     "crypto": false
+  },
+  "devDependencies": {
+    "@babel/preset-typescript": "^7.24.1",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.70",
+    "@types/react-dom": "^18.2.22",
+    "@types/react-modal": "^3.16.3",
+    "typescript": "^5.4.3"
   }
 }
