+++ src/Components/common/responsive.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import styled from 'styled-components';
 
 const ResponsiveBlock = styled.div`
@@ -16,7 +16,11 @@ const ResponsiveBlock = styled.div`
   }
 `;
 
-const Responsive = ({ children, ...rest }) => {
+interface ResponsiveProps {
+  children: ReactNode;
+}
+
+const Responsive = ({ children, ...rest }: ResponsiveProps) => {
   // style, className, onClick, onMouseMove 등의 props를 사용할 수 있도록
   // ...rest를 사용하여 ResponsiveBlock에게 전달
   return <ResponsiveBlock {...rest}>{children}</ResponsiveBlock>;
