+++ src/Pages/InternalPage/Dashboard/Today.tsx
@@ -14,9 +14,8 @@ const Subtitle = styled.text`
 const ListItem = styled.div`
   margin-top: 5px;
 `;
-function Today({ projectId }) {
-  const [todayEvents, setTodayEvents] = useState([]);
-  const [message, setMessage] = useState("");
+function Today({ projectId }: { projectId: number }) { // projectId 타입 명시
+  const [todayEvents, setTodayEvents] = useState<any[]>([]); // todayEvents의 타입 명시
 
   useEffect(() => {
     const fetchTodayEvents = async () => {
@@ -29,7 +28,7 @@ function Today({ projectId }) {
         const response = await scheduleApi.getScheduleList(projectId);
         if (response.data && response.data.success === false) {
           if (response.data.code === 6001) {
-            setMessage(response.data.message); // "일정 목록이 존재하지 않습니다."
+            console.log(response.data.message); // "일정 목록이 존재하지 않습니다."
           } else if (response.data.code === 7001) {
             sessionStorage.removeItem("login-token");
             delete axios.defaults.headers.common["Authorization"];
@@ -37,14 +36,14 @@ function Today({ projectId }) {
           }
           return;
         }
-        const filteredEvents = response.data.list.filter((e) => {
+        const filteredEvents = response.data.list.filter((e: any) => { // e의 타입 명시
           const eventStartDateTimestamp = new Date(e.startDate).getTime();
           const eventEndDateTimestamp = new Date(e.endDate).getTime();
 
           // startDate가 오늘이거나, 오늘이 startDate와 endDate 사이에 있거나, endDate가 오늘인 경우를 모두 고려
           return (
-            todayTimestamp >= eventStartDateTimestamp &&
-            todayTimestamp <= eventEndDateTimestamp
+              todayTimestamp >= eventStartDateTimestamp &&
+              todayTimestamp <= eventEndDateTimestamp
           );
         });
 
@@ -58,18 +57,18 @@ function Today({ projectId }) {
   }, [projectId]);
 
   return (
-    <Container>
-      <Subtitle>Today</Subtitle>
-      {todayEvents.length === 0 ? (
-        <p>오늘의 일정이 없습니다.</p>
-      ) : (
-        todayEvents.map((event) => (
-          <ListItem key={event.startDate}>
-            <li>{event.content}</li>
-          </ListItem>
-        ))
-      )}
-    </Container>
+      <Container>
+        <Subtitle>Today</Subtitle>
+        {todayEvents.length === 0 ? (
+            <p>오늘의 일정이 없습니다.</p>
+        ) : (
+            todayEvents.map((event) => (
+                <ListItem key={event.startDate}>
+                  <li>{event.content}</li>
+                </ListItem>
+            ))
+        )}
+      </Container>
   );
 }
