+++ src/Pages/index.js
@@ -1,5 +1,5 @@
 import {Login} from "./Login/login.js";
 import {Portfolio} from "./Portfolio/portfolio.js";
-import {AboutUs} from "./AboutUs/about-us.js";
+import {AboutUs} from "./AboutUs/about-us.tsx";
 
 export {Login, Portfolio, AboutUs};
