+++ src/api/checkTodoApi.tsx
@@ -1,48 +1,45 @@
-import axios from "axios";
-const storedToken = sessionStorage.getItem('login-token');
+import axios, { AxiosResponse } from "axios";
+
+const storedToken: string | null = sessionStorage.getItem('login-token');
 if (storedToken) {
   axios.defaults.headers.common['Authorization'] = 'Bearer ' + storedToken;
 }
-// 체크리스트 crud api
-const checkTodoApi = {
-  // 할일 선택 조회 (없애도됌)
+
+interface CheckTodoApi {
+  getCheckTodo: (todoIndex: number) => Promise<AxiosResponse>;
+  updateCheckTodo: (todoIndex: number, data: { todoContent: string; todoEmergency: boolean }) => Promise<AxiosResponse>;
+  deleteCheckTodo: (todoIndex: number) => Promise<AxiosResponse>;
+  getProjectTodo: (projectId: number) => Promise<AxiosResponse>;
+  createCheckTodo: (projectId: number, data: { todoContent: string; todoEmergency: boolean }) => Promise<AxiosResponse>;
+  pingTest: () => Promise<AxiosResponse>;
+  completeCheckTodo: (todoIndex: number) => Promise<AxiosResponse>;
+}
+
+const checkTodoApi: CheckTodoApi = {
   getCheckTodo: async (todoIndex) => {
     const response = await axios.get(`/api/todo/${todoIndex}`);
     return response;
   },
-  // 할일 업데이트
-  // data안에는 객체로 todoContent, todoEmergency가들어있어야 함.(없애도됌)
   updateCheckTodo: async (todoIndex, data) => {
     const response = await axios.put(`/api/todo/${todoIndex}`, data);
     return response;
   },
-  // 할일 삭제
   deleteCheckTodo: async (todoIndex) => {
     const response = await axios.delete(`/api/todo/${todoIndex}`);
     return response;
   },
-  // 할일 조회
   getProjectTodo: async (projectId) => {
-    const response = await axios.get(
-      `/api/projects/${projectId}/todo`
-    );
+    const response = await axios.get(`/api/projects/${projectId}/todo`);
     return response;
   },
-  // 할일 등록
-  // data안에는 객체로 todoContent, todoEmergency가들어있어야 함.
   createCheckTodo: async (projectId, data) => {
-    const response = await axios.post(
-      `/api/projects/${projectId}/todo`,
-      data
-    );
+    const response = await axios.post(`/api/projects/${projectId}/todo`, data);
     return response;
   },
-  // ping test (이건 안해도 괜춘)
   pingTest: async () => {
     const response = await axios.get(`/api/todo/ping`);
     return response;
   },
-  // 할일 완료 표시
   completeCheckTodo: async (todoIndex) => {
     const response = await axios.get(`/api/todo/finish/${todoIndex}`);
     return response;
