+++ src/assets/logo/images.d.ts
@@ -1,0 +1,3 @@
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.jpeg';
